refactor: implemented review comments
